requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
python-dotenv
psycopg2-binary
